Support sidecar manfiest files, add cryptomatte manifest driver for writing said files, split cryptomatte.h into cryptomatte.h and cryptomatte.cpp, changed calling conventions to use prototyped funcs
