+++ cryptomatte/CMakeLists.txt
@@ -1,7 +1,8 @@
 install(DIRECTORY nuke DESTINATION ${INSTALL_DIR})
 
+
+set(SRC cryptomatte.cpp cryptomatte_loader.cpp cryptomatte_shader.cpp cryptomatte_manifest_driver.cpp cryptomatte_filter.cpp MurmurHash3.cpp )
 set(SHADER cryptomatte)
-set(SRC cryptomatte_loader.cpp ${SHADER}.cpp cryptomatte_filter.cpp MurmurHash3.cpp)
 set(UI ${CMAKE_CURRENT_SOURCE_DIR}/${SHADER}.ui)
 set(MTD ${CMAKE_CURRENT_BINARY_DIR}/${SHADER}.mtd)
 set(AE ${CMAKE_CURRENT_BINARY_DIR}/cryptomatteTemplate.py)
